Move ISection interface to its own directory
