+++ src/interfaces/Section/index.ts
@@ -1,0 +1,6 @@
+export interface ISection {
+  uuid?: string;
+  title: string;
+  text: string;
+  displayOrder: number;
+}
